requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Text Processing
nltk>=3.8
spacy>=3.6.0
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
pyarrow>=12.0.0
h5py>=3.8.0

# API and Web
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=0.27.0
httpx>=0.24.0

# Configuration and Logging
pyyaml>=6.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Progress and Visualization
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Rich text output (optional)
rich>=13.4.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
click>=8.1.0
pathlib2>=2.3.0
filelock>=3.12.0
psutil>=5.9.0

# Specific versions for compatibility
torch-audio>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# Optional: CUDA support
# torch-cuda>=11.8  # Uncomment if using CUDA 11.8
# torch-cuda>=12.1  # Uncomment if using CUDA 12.1

# For distributed training (optional)
deepspeed>=0.9.0

# For model serving (optional)
fastapi>=0.100.0
uvicorn>=0.22.0